所有rebar band,加入菜单,隐藏显示Toobar中. 修正rebar高度更改时,更新页面布局.
git-svn-id: svn://172.16.0.3@76 c573b714-58c8-aa40-881b-c130d9d1abad

+++ playerT/PlayerThread.cpp
@@ -84,7 +84,7 @@ void CPlayerThread::WriteDataToDSBuf()
 
 	DOUBLE used,lefted;
 	m_pPlayer->m_pFile->GetPos(&used,&lefted);
-	::PostMessage(m_pPlayer->m_pMainFrame->m_hWnd,WM_TRACKPOS,used,lefted);
+	::PostMessage(m_pPlayer->m_pMainFrame->m_hWnd,WM_TRACKPOS,(int)used,(int)lefted);
 	
 	DWORD playCursor;
 	if (FAILED(m_lpDSBuffer->GetCurrentPosition(&playCursor,NULL))) return;
